call jwt auth with sdk only at conference page
